# firmware/openapi.yaml
openapi: "3.0.3"
info:
  title: Arbor Node API
  description: |
    REST API exposed by the Arbor ESP32 firmware.
    This is the on-device API — the control server proxies these endpoints.
    Authentication uses Bearer tokens with SHA-256 hashed API keys.
  version: "0.1.0"
  contact:
    name: Arbor
  license:
    name: Proprietary
servers:
  - url: https://{host}:{port}/api/v1
    description: Node HTTPS (default)
    variables:
      host:
        default: arbor.local
      port:
        default: "8443"
  - url: http://{host}:{port}/api/v1
    description: Node HTTP (provisioning/dev)
    variables:
      host:
        default: arbor.local
      port:
        default: "8443"
...
security:
  - BearerAuth: []
paths:
...
  /system/config:
    get:
      tags: [system]
      summary: Get current configuration
      description: Returns full config with secrets redacted. Requires READ scope.
      responses:
        "200":
          description: Current configuration
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [system]
      summary: Update configuration
      description: Validates, saves to NVS, updates running config. Requires ADMIN scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                  restart_required:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /system/restart:
    post:
      tags: [system]
      summary: Restart the node
      description: Triggers esp_restart() after sending response. Requires ADMIN scope.
      responses:
        "200":
          description: Restarting
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key as Bearer token. Keys are SHA-256 hashed with per-key salt
        and stored in NVS. In provisioning mode (no keys loaded), all
        requests are allowed.
...
  responses:
    NotFound:
      description: Not found (also used for unauthorized per SECURITY.md)
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...